Add resolution queries and types files
and remove commented out code from the scopes impl

---
