+++ compiler/next/include/chpl/resolution/resolution-queries.h
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2021 Hewlett Packard Enterprise Development LP
+ * Other additional copyright holders may be indicated within.
+ *
+ * The entirety of this work is licensed under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ *
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef CHPL_RESOLUTION_RESOLUTION_QUERIES_H
+#define CHPL_RESOLUTION_RESOLUTION_QUERIES_H
+
+#include "chpl/resolution/resolution-types.h"
+
+namespace chpl {
+namespace resolution {
+
+
+  /*
+  // Resolves the top-level declarations in a module
+  const ResolvedSymbol&
+  resolveModule(Context* context, const uast::Module* mod);
+
+  const ResolvedSymbolVec& resolveFile(Context* context, UniqueString path);
+   */
+
+
+} // end namespace resolution
+} // end namespace chpl
+#endif
