+++ compiler/next/lib/resolution/CMakeLists.txt
@@ -18,6 +18,8 @@
 target_sources(libchplcomp-obj
                PRIVATE
 
+               resolution-queries.cpp
+               resolution-types.cpp
                scope-queries.cpp
                scope-types.cpp
